review: apply suggestions + remove cpu backend flag
