+++ optimum/neuron/modeling_diffusion.py
@@ -63,6 +63,7 @@ from .utils import (
     NeuronArgumentParser,
     check_if_weights_replacable,
     get_neuron_instance_type,
+    is_cpu_only_instance,
     replace_weights,
     store_compilation_config,
 )
@@ -791,7 +792,6 @@ class NeuronDiffusionPipelineBase(NeuronTracedModel):
         disable_neuron_cache: bool = False,
         inline_weights_to_neff: bool = True,
         instance_type: Literal["trn1", "inf2", "trn1n", "trn2"] | None = None,
-        cpu_backend: bool = False,
         optlevel: str = "2",
         subfolder: str = "",
         local_files_only: bool = False,
@@ -905,6 +905,7 @@ class NeuronDiffusionPipelineBase(NeuronTracedModel):
 
         # Get compilation arguments
         auto_cast_type = None if auto_cast is None else auto_cast_type
+        cpu_backend = is_cpu_only_instance()
         instance_type = get_neuron_instance_type(instance_type)
         compiler_kwargs = {
             "auto_cast": auto_cast,
@@ -983,7 +984,6 @@ class NeuronDiffusionPipelineBase(NeuronTracedModel):
                     compiler_version=NEURON_COMPILER_VERSION,
                     inline_weights_to_neff=inline_weights_to_neff,
                     optlevel=optlevel,
-                    cpu_backend=cpu_backend,
                     model_type=getattr(neuron_config, "MODEL_TYPE", None),
                     task=getattr(neuron_config, "task", None),
                     output_attentions=output_attentions,
