+++ optimum/neuron/utils/__init__.py
@@ -89,7 +89,7 @@ _import_structure = {
         "is_model_officially_supported",
         "patch_transformers_for_neuron_sdk",
     ],
-    "system": ["SUPPORTED_INSTANCE_TYPES", "get_neuron_instance_type"],
+    "instance": ["SUPPORTED_INSTANCE_TYPES", "get_neuron_instance_type", "is_cpu_only_instance"],
 }
 
 if TYPE_CHECKING:
@@ -136,6 +136,7 @@ if TYPE_CHECKING:
         DummyTimestepInputGenerator,
         WhisperDummyTextInputGenerator,
     )
+    from .instance import SUPPORTED_INSTANCE_TYPES, get_neuron_instance_type, is_cpu_only_instance
     from .misc import (
         DiffusersPretrainedConfig,
         check_if_weights_replacable,
@@ -159,7 +160,6 @@ if TYPE_CHECKING:
         patch_within_function,
         replace_class_in_inheritance_hierarchy,
     )
-    from .system import SUPPORTED_INSTANCE_TYPES, get_neuron_instance_type
     from .training_utils import (
         is_model_officially_supported,
         patch_transformers_for_neuron_sdk,
